+++ mayan/apps/file_metadata/icons.py
@@ -1,10 +1,16 @@
 from mayan.apps.appearance.classes import Icon
 
 icon_document_submit = Icon(
-    driver_name='fontawesome', symbol='chess-board'
+    driver_name='fontawesome-dual', primary_symbol='chess-board',
+    secondary_symbol='arrow-right'
 )
 icon_document_multiple_submit = Icon(
-    driver_name='fontawesome', symbol='chess-board'
+    driver_name='fontawesome-dual', primary_symbol='chess-board',
+    secondary_symbol='arrow-right'
+)
+icon_document_type_submit = Icon(
+    driver_name='fontawesome-dual', primary_symbol='chess-board',
+    secondary_symbol='arrow-right'
 )
 icon_file_metadata = Icon(
     driver_name='fontawesome', symbol='chess-board'