+++ mayan/apps/ocr/icons.py
@@ -2,22 +2,32 @@ from mayan.apps.appearance.classes import Icon
 
 icon_document_ocr_content = Icon(driver_name='fontawesome', symbol='font')
 icon_document_ocr_content_delete = Icon(
-    driver_name='fontawesome-dual',
-    primary_symbol='font',
-    secondary_symbol='minus'
+    driver_name='fontawesome-dual', primary_symbol='font',
+    secondary_symbol='times'
 )
 icon_document_page_ocr_content = Icon(driver_name='fontawesome', symbol='font')
 icon_document_multiple_submit = Icon(driver_name='fontawesome', symbol='font')
 icon_document_ocr_download = Icon(
-    driver_name='fontawesome', symbol='font'
+    driver_name='fontawesome-dual', primary_symbol='font',
+    secondary_symbol='arrow-down'
 )
 icon_document_ocr_errors_list = Icon(
-    driver_name='fontawesome', symbol='font'
+    driver_name='fontawesome-dual', primary_symbol='font',
+    secondary_symbol='exclamation'
 )
 icon_document_type_ocr_settings = Icon(
     driver_name='fontawesome', symbol='font'
 )
-icon_document_type_submit = Icon(driver_name='fontawesome', symbol='font')
-icon_entry_list = Icon(driver_name='fontawesome', symbol='font')
-
+icon_document_type_submit = Icon(
+    driver_name='fontawesome-dual', primary_symbol='font',
+    secondary_symbol='arrow-right'
+)
+icon_document_multiple_submit = Icon(
+    driver_name='fontawesome-dual', primary_symbol='font',
+    secondary_symbol='arrow-right'
+)
+icon_entry_list = Icon(
+    driver_name='fontawesome-dual', primary_symbol='font',
+    secondary_symbol='exclamation'
+)
 icon_document_submit = icon_document_multiple_submit