+++ mayan/apps/dependencies/dependencies.py
@@ -1,5 +1,5 @@
 from .classes import PythonDependency
 
 PythonDependency(
-    module=__name__, name='node-semver', version_string='==0.6.1'
+    module=__name__, name='node-semver', version_string='==0.8.0'
 )