+++ mayan/apps/metadata/dependencies.py
@@ -35,5 +35,5 @@ PythonDependency(
         LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
         NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
         SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-    ''', module=__name__, name='python-dateutil', version_string='==2.8.0'
+    ''', module=__name__, name='python-dateutil', version_string='==2.8.1'
 )